don't try to remove a AccountLoginSession if the account isn't logged in. this can happen (rarely) when we use logout() to clear registered session objects.
svn commit r77303

+++ Site/SiteAccountSessionModule.php
@@ -532,14 +532,16 @@ class SiteAccountSessionModule extends SiteSessionModule
 
 	public function unsetLoginSession()
 	{
-		$sql = sprintf(
-			'delete from AccountLoginSession
-			where session_id = %s and account = %s',
-			$this->app->db->quote($this->getSessionId(), 'text'),
-			$this->app->db->quote($this->account->id, 'integer')
-		);
+		if ($this->isLoggedIn()) {
+			$sql = sprintf(
+				'delete from AccountLoginSession
+				where session_id = %s and account = %s',
+				$this->app->db->quote($this->getSessionId(), 'text'),
+				$this->app->db->quote($this->account->id, 'integer')
+			);
 
-		SwatDB::exec($this->app->db, $sql);
+			SwatDB::exec($this->app->db, $sql);
+		}
 	}
 
 	// }}}
